moves movie functionality to movies.js
